Improve scraping logic to ensure finding element
-Also, vary one of the tests to use another office choice.

Time waits are made to two seconds to guarantee that the elements are
found. While explicit waits could have been used instead of implicit
waits, the code would have been less readable. I choose readability
since anyway the task is running in the background, so the performance
tradeoff will not compromise usability of the tool.